+++ src/oracle_tests/mod.rs
@@ -1,4 +1,5 @@
 mod funnel_leaf;
+mod invalid_leaf;
 mod macro_leaf;
 mod sequence;
 mod span;